libmtl: drop lib prefix, rename to 'mtl'
