--- src/libmtl/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(libmtl SHARED MTLFormula.cpp)
-target_include_directories(libmtl PUBLIC include)
