+++ src/mtl/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(mtl SHARED MTLFormula.cpp)
+target_include_directories(mtl PUBLIC include)
