+++ test/CMakeLists.txt
@@ -7,11 +7,11 @@ target_link_libraries(testta PRIVATE libta PRIVATE Catch2::Catch2)
 catch_discover_tests(testta)
 
 add_executable(testinterval test_interval.cpp catch2_main.cpp)
-target_link_libraries(testinterval PRIVATE libmtl PRIVATE Catch2::Catch2)
+target_link_libraries(testinterval PRIVATE mtl PRIVATE Catch2::Catch2)
 catch_discover_tests(testinterval)
 
 add_executable(testmtlformulae test_mtlFormula.cpp catch2_main.cpp)
-target_link_libraries(testmtlformulae PRIVATE libmtl PRIVATE Catch2::Catch2)
+target_link_libraries(testmtlformulae PRIVATE mtl PRIVATE Catch2::Catch2)
 catch_discover_tests(testmtlformulae)
 
 add_executable(testata catch2_main.cpp test_ata_formula.cpp test_ata.cpp)
