+++ test/test_mtlFormula.cpp
@@ -18,8 +18,8 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
-#include <libmtl/MTLFormula.h>
-#include <libmtl/operators.h>
+#include <mtl/MTLFormula.h>
+#include <mtl/operators.h>
 
 #include <catch2/catch.hpp>
